Move holidays to their own directory
