+++ pandas_market_calendars/exchange_calendars/cme_globex_agriculture.py
@@ -13,22 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from abc import ABC, abstractmethod
+from abc import abstractmethod
 from .cme_globex_base import CMEGlobexBaseExchangeCalendar
 
 from datetime import time
-from itertools import chain
 
-from pandas import Timestamp
 from pandas.tseries.holiday import AbstractHolidayCalendar, GoodFriday, USLaborDay, USPresidentsDay, USThanksgivingDay
-from pytz import timezone
 
-from pandas_market_calendars.holidays_us import (Christmas, ChristmasEveBefore1993, ChristmasEveInOrAfter1993, USBlackFridayInOrAfter1993,
-                                                 USIndependenceDay, USMartinLutherKingJrAfter1998, USMemorialDay, USJuneteenthAfter2022,
-                                                 USNationalDaysofMourning, USNewYearsDay)
-from pandas_market_calendars.market_calendar import MarketCalendar
-
-
+from pandas_market_calendars.holidays.us import (Christmas, ChristmasEveBefore1993, ChristmasEveInOrAfter1993, USBlackFridayInOrAfter1993,
+                                                 USIndependenceDay, USMartinLutherKingJrAfter1998, USMemorialDay,
+                                                 USNewYearsDay)
 
 
 class CMEGlobexAgricultureExchangeCalendar(CMEGlobexBaseExchangeCalendar):
