+++ pandas_market_calendars/exchange_calendars/cme_globex_base.py
@@ -15,16 +15,12 @@
 
 from abc import ABC, abstractmethod
 
-from datetime import time
-from itertools import chain
-
-from pandas import Timestamp
 from pandas.tseries.holiday import AbstractHolidayCalendar, GoodFriday, USLaborDay, USPresidentsDay, USThanksgivingDay
 from pytz import timezone
 
-from pandas_market_calendars.holidays_us import (Christmas, ChristmasEveBefore1993, ChristmasEveInOrAfter1993, USBlackFridayInOrAfter1993,
+from pandas_market_calendars.holidays.us import (Christmas, ChristmasEveBefore1993, ChristmasEveInOrAfter1993, USBlackFridayInOrAfter1993,
                                                  USIndependenceDay, USMartinLutherKingJrAfter1998, USMemorialDay, USJuneteenthAfter2022,
-                                                 USNationalDaysofMourning, USNewYearsDay)
+                                                 USNewYearsDay)
 from pandas_market_calendars.market_calendar import MarketCalendar
