+++ pandas_market_calendars/exchange_calendars/cme_globex_energy_and_metals.py
@@ -13,14 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from abc import ABC, abstractmethod
 from .cme_globex_base import CMEGlobexBaseExchangeCalendar
 
 from datetime import time
-from itertools import chain
 
-import pandas as pd
-from pandas import Timestamp
 from pandas.tseries.holiday import AbstractHolidayCalendar #, GoodFriday, USLaborDay, USPresidentsDay, USThanksgivingDay
 from pytz import timezone
 
@@ -28,8 +24,7 @@ from pytz import timezone
 #                          USIndependenceDay, USMartinLutherKingJrAfter1998, USMemorialDay, USJuneteenthAfter2022,
 #                          USNationalDaysofMourning, USNewYearsDay)
 
-from pandas_market_calendars.holidays_cme_globex import (USNewYearsDay,
-                                                         USMartinLutherKingJrFrom2022, USMartinLutherKingJrPre2022, USNewYearsDay,
+from pandas_market_calendars.holidays.cme_globex import (USMartinLutherKingJrFrom2022, USMartinLutherKingJrPre2022, USNewYearsDay,
                                                          USPresidentsDayFrom2022, USPresidentsDayPre2022,
                                                          GoodFriday,
                                                          USMemorialDayFrom2022, USMemorialDayPre2022,
@@ -38,9 +33,6 @@ from pandas_market_calendars.holidays_cme_globex import (USNewYearsDay,
                                                          USLaborDay,
                                                          USThanksgivingDayFrom2022, USThanksgivingDayPre2022, FridayAfterThanksgiving,
                                                          ChristmasCME)
-from pandas_market_calendars.market_calendar import MarketCalendar
-
-
 
 
 class CMEGlobexEnergyAndMetalsExchangeCalendar(CMEGlobexBaseExchangeCalendar):
