+++ pandas_market_calendars/exchange_calendars/cme_globex_fixed_income.py
@@ -2,9 +2,8 @@ from .cme_globex_base import CMEGlobexBaseExchangeCalendar
 
 from datetime import time
 from pandas.tseries.holiday import AbstractHolidayCalendar
-from pytz import timezone
 
-from pandas_market_calendars.holidays_cme import (
+from pandas_market_calendars.holidays.cme import (
     USMartinLutherKingJrAfter1998Before2015,
     USMartinLutherKingJrAfter1998Before2016FridayBefore,
     USMartinLutherKingJrAfter2015,
@@ -30,7 +29,7 @@ from pandas_market_calendars.holidays_cme import (
     USThanksgivingAfter2014,
     USThanksgivingFriday,
 )
-from pandas_market_calendars.holidays_us import (
+from pandas_market_calendars.holidays.us import (
     USNewYearsDay,
     ChristmasEveInOrAfter1993,
     Christmas,
