+++ pandas_market_calendars/exchange_calendars/iex.py
@@ -4,7 +4,7 @@ from .nyse import NYSEExchangeCalendar
 from pandas.tseries.holiday import AbstractHolidayCalendar
 from pytz import timezone 
 
-from pandas_market_calendars.holidays_nyse import (
+from pandas_market_calendars.holidays.nyse import (
     USPresidentsDay,
     GoodFriday,
     USMemorialDay,
