+++ pandas_market_calendars/exchange_calendars/jpx.py
@@ -4,8 +4,8 @@ from itertools import chain
 from pandas.tseries.holiday import AbstractHolidayCalendar
 from pytz import timezone
 
-from pandas_market_calendars.holidays_jp import *
-from pandas_market_calendars.holidays_us import USNewYearsDay
+from pandas_market_calendars.holidays.jp import *
+from pandas_market_calendars.holidays.us import USNewYearsDay
 from pandas_market_calendars.market_calendar import MarketCalendar
