+++ pandas_market_calendars/exchange_calendars/lse.py
@@ -18,7 +18,7 @@ from datetime import time
 from pandas.tseries.holiday import AbstractHolidayCalendar, EasterMonday, GoodFriday
 from pytz import timezone
 
-from pandas_market_calendars.holidays_uk import (
+from pandas_market_calendars.holidays.uk import (
     BoxingDay, Christmas, ChristmasEve, LSENewYearsDay, LSENewYearsEve,
     MayBank_pre_1995, MayBank_post_1995_pre_2020, MayBank_post_2020,
     SpringBank_pre_2002, SpringBank_post_2002_pre_2012, SpringBank_post_2012_pre_2022, SpringBank_post_2022,
