+++ pandas_market_calendars/exchange_calendars/sifma.py
@@ -3,8 +3,6 @@ from datetime import time
 from pandas.tseries.holiday import AbstractHolidayCalendar
 from pytz import timezone
 from itertools import chain
-import pandas as pd
-
 
 ########################################################################################################################
 # SIFMA Financial Markets Calendar for US, UK, JP
@@ -22,9 +20,8 @@ import pandas as pd
 ########################################################################################################################
 
 
-from pandas_market_calendars.holidays_sifma import(
+from pandas_market_calendars.holidays.sifma import(
     # US Holidays
-    UKWeekendBoxingDay,
     USNewYearsDay,  # Not observed if a Saturday
     USNewYearsEve2pmEarlyClose,
 
@@ -214,7 +211,7 @@ class SIFMAUKExchangeCalendar(MarketCalendar):
 ############################################################
 # Japan
 ############################################################
-from pandas_market_calendars.holidays_jp import (JapanComingOfAgeDay, JapanNationalFoundationDay, JapanEmperorsBirthday, JapanVernalEquinox, JapanShowaDay,
+from pandas_market_calendars.holidays.jp import (JapanComingOfAgeDay, JapanNationalFoundationDay, JapanEmperorsBirthday, JapanVernalEquinox, JapanShowaDay,
                                                  JapanConstitutionMemorialDay, JapanGreeneryDay, JapanChildrensDay, JapanMarineDay, JapanMountainDay,
                                                  JapanRespectForTheAgedDay, JapanAutumnalEquinox,
                                                  JapanHealthAndSportsDay2000To2019, JapanSportsDay2020, JapanSportsDay,
