+++ pandas_market_calendars/exchange_calendars/sse.py
@@ -4,7 +4,7 @@ from functools import partial
 from pandas.tseries.holiday import AbstractHolidayCalendar, Holiday, next_monday
 from pytz import timezone
 
-from pandas_market_calendars.holidays_cn import *
+from pandas_market_calendars.holidays.cn import *
 from pandas_market_calendars.market_calendar import MarketCalendar
