+++ pandas_market_calendars/holidays/cme.py
