+++ pandas_market_calendars/holidays/jp.py
@@ -2,7 +2,7 @@ from dateutil.relativedelta import MO
 from pandas import DateOffset, Timestamp
 from pandas.tseries.holiday import Holiday, sunday_to_monday
 
-from pandas_market_calendars.jpx_equinox import autumnal_citizen_dates, autumnal_equinox, vernal_equinox
+from pandas_market_calendars.holidays.jpx_equinox import autumnal_citizen_dates, autumnal_equinox, vernal_equinox
 
 AscensionDays = [
     Timestamp('2019-04-30', tz='UTC'),  # National Holiday
